getMuPar {macroutils}	R Documentation

Get a single default parameters for the package.

Description

Get a single default parameters for the package. Wrapper around muPar.

Usage

getMuPar(par)

Arguments

par
See the par argument in muPar. Notice that if more than one parameter name is provided, only the first one will be returned.

Value

Returns the value of the parameter par, without the list container of muPar.

[Package macroutils version 1.14.0 Index]