macroBugFixCleanDb {macroutils}	R Documentation

Clean-up MACRO 5.2 parameter databases. Fixes 4 known bugs (orphan, incomplete or unnecessary values)

Description

Clean-up MACRO 5.2 parameter databases. Fixes 4 known bugs (orphan, incomplete or unnecessary values). It is very highly recommended to make a backup-copy of MACRO 5.2 parameter databases before you try this utility. The R RODBC-package is required to run this function, and you also need to run a 32 bit (i386) version of R (maybe located in {R installation directory}bin/i386/Rgui.exe, if it has been installed).

Usage

macroBugFixCleanDb(file, paranoia = TRUE, ...)

Arguments

file
Vector of character strings or a single connection to a MACRO GUI MS Access parameter database. If a vector of character strings, it should be the name(s) of the Access database(s) containing MACRO parameters. The path of the file(s) may be provided as well, if file(s) is (are) not in the working directory.

paranoia
Single logical value. If TRUE, the user is asked if he made a backup copy of the parameter database.

...
Additional options passed to specific methods.

Value

Do not return anything. Used for it side effect on a MACRO parameter database.

[Package macroutils version 1.12.0 Index]