macroReadIndump {macroutils}	R Documentation

INTERNAL. Import a MACRO indump.tmp file and output it in a human readable format.

Description

INTERNAL. Import a MACRO indump.tmp file and output it in a human readable format. It reads layered parameters, options, crop parameters and irrigation parameters, but not yet output parameters. EXPERIMENTAL. USE AT YOUR OWN RISKS.

Usage

macroReadIndump(f, layerLoc = 7, exportTrash = FALSE)

Arguments

f
Single character string. Name (and if needed, path) of the indump.tmp file to be read

layerLoc
Single integer. Line where the number of numerical layers is written

exportTrash
Single logical value. If TRUE, 'filling' parameter values (i.e. values written but not used) are also exported.

Value

Returns a list of data.frames with different MACRO parameters

[Package macroutils version 1.12.0 Index]