macroViewBin {macroutils}	R Documentation

Reads a MACRO/SOIL binary file and view it as a table.

Description

Reads a MACRO/SOIL binary file and view it as a table.

Usage

macroViewBin(file, ...)

Arguments

file
Single character strings or a single connection to a binary file. If a vector character strings, it should be the name of the binary file which the data are to be read from. The path of the file may be provided as well, if file is not in the working directory.

...
More arguments passed to macroReadBin.

[Package macroutils version 1.14.0 Index]