macroWriteBin {macroutils}	R Documentation

Write bin file for the SOIL and MACRO models.

Description

Write bin file for the SOIL and MACRO models. Origial code by Kristian Persson. R code vectorisation by Julien Moeys.

Usage

macroWriteBin(x, ...)

## S3 method for class 'macroTimeSeries'
macroWriteBin(x, ...)

## S3 method for class 'macroTimeSeriesList'
macroWriteBin(x, file, ...)

## S3 method for class 'list'
macroWriteBin(x, ...)

## S3 method for class 'data.frame'
macroWriteBin(x, file, ...)

Arguments

x
A data.frame. Table of data to be written in file. The table must contain one column named "Date" containing POSIXct dates, and thus must have column names. All columns but "Date" must be of type numerical (integer or double), and will be written as double. The "Date" will be converted into integers, representing minutes since 2 days before the 1st of Januray of year 0001 at 00:00. Missing values are not allowed.

file
Single character string or connection to a binary file. If a character string, it should be the name of the binary file which the data are to be written from. The path of the file may be provided as well, if file is not in the working directory.

...
Additional options passed to writeBin

Examples

library( "macroutils" )

# ====== Write a binary file in SOIL or MACRO style ======

# 1.1. Read a table, that will then be written back

#   Path to the file to be read
( filenm <- system.file( "bintest/METFILE.BIN", 
    package = "macroutils", mustWork = TRUE ) )

## [1] "C:/Users/*******/Documents/R/win-library/3.3/macroutils/bintest/METFILE.BIN"

#   Read the file
tmp1 <- macroReadBin( file = filenm ) 

# 1.2. Generate a dummy temporary file where the table will be 
#   written

( filenm <- tempfile(  ) )

## [1] "C:\\Users\\*******\\AppData\\Local\\Temp\\Rtmp6ZAx3P\\file1394427449ea"

# 1.3. Write this table in SOIL or MACRO bin style 

#   NB: This table is NOT a standard SOIL or MACRO input file!

macroWriteBin( file = filenm, x = tmp1 ) 

#   NB: When writing the bin file, time zones are ignored!

# 1.4. Read that file again and check that it is the same:

tmp1.b <- macroReadBin( file = filenm ) 

# Maximum difference (small numerical differences)
unlist( lapply(
    X      = colnames(tmp1), 
    FUN    = function(X){ 
        max( tmp1[,X] - tmp1.b[,X] ) 
    }   
) ) 

## [1] 0 0 0 0 0 0

# ====== Miscellaneous (developers niche) ======

#   Comparison with the initial implementation of the R code 
#   Note: programmers playground

filenm2 <- tempfile(  ) 
filenm2 

## [1] "C:\\Users\\*******\\AppData\\Local\\Temp\\Rtmp6ZAx3P\\file1394ced4146"

#   Write the table a second time
macroutils:::.write.BIN.old( 
    filename = filenm2, 
    data.to.write = tmp1 )   

## [1] "Date"                                       
## [2] "Max_air_temperature_oC_Sutton_Bonington"    
## [3] "Min_air_temperature_oC_Sutton_Bonington"    
## [4] "Solar_radiation_W_m_2_Sutton_Bonington_TRS1"
## [5] "Vapour_pressure_kPa_Sutton_Bonington_TRS1"  
## [6] "Wind_speed_m_s_1_Sutton_Bonington_TRS1"     
## [1] "Max_air_temperature_oC_Sutton_Bonington"
## [1] "Min_air_temperature_oC_Sutton_Bonington"
## [1] "Solar_radiation_W_m_2_Sutton_Bonington_TRS1"
## [1] "Vapour_pressure_kPa_Sutton_Bonington_TRS1"
## [1] "Wind_speed_m_s_1_Sutton_Bonington_TRS1"

#   Read the table a second time:
tmp1.c <- macroutils:::.read.BIN.old( 
    filename = filenm2 ) 

#   Compare the 2 file written / read:
identical( tmp1.b, tmp1.c ) 

## [1] FALSE

data.frame( "A"=tmp1[,1], "B"=tmp1.b[,1], "C"=tmp1.c[,1] )

##                       A                   B                   C
## 1   1987-01-01 12:00:00 1987-01-01 12:00:00 1987-01-01 12:00:00
## 2   1987-01-02 12:00:00 1987-01-02 12:00:00 1987-01-02 12:00:00
## 3   1987-01-03 12:00:00 1987-01-03 12:00:00 1987-01-03 12:00:00
## 4   1987-01-04 12:00:00 1987-01-04 12:00:00 1987-01-04 12:00:00
## 5   1987-01-05 12:00:00 1987-01-05 12:00:00 1987-01-05 12:00:00
## 6   1987-01-06 12:00:00 1987-01-06 12:00:00 1987-01-06 12:00:00
## 7   1987-01-07 12:00:00 1987-01-07 12:00:00 1987-01-07 12:00:00
## 8   1987-01-08 12:00:00 1987-01-08 12:00:00 1987-01-08 12:00:00
## 9   1987-01-09 12:00:00 1987-01-09 12:00:00 1987-01-09 12:00:00
## 10  1987-01-10 12:00:00 1987-01-10 12:00:00 1987-01-10 12:00:00
## 11  1987-01-11 12:00:00 1987-01-11 12:00:00 1987-01-11 12:00:00
## 12  1987-01-12 12:00:00 1987-01-12 12:00:00 1987-01-12 12:00:00
## 13  1987-01-13 12:00:00 1987-01-13 12:00:00 1987-01-13 12:00:00
## 14  1987-01-14 12:00:00 1987-01-14 12:00:00 1987-01-14 12:00:00
## 15  1987-01-15 12:00:00 1987-01-15 12:00:00 1987-01-15 12:00:00
## 16  1987-01-16 12:00:00 1987-01-16 12:00:00 1987-01-16 12:00:00
## 17  1987-01-17 12:00:00 1987-01-17 12:00:00 1987-01-17 12:00:00
## 18  1987-01-18 12:00:00 1987-01-18 12:00:00 1987-01-18 12:00:00
## 19  1987-01-19 12:00:00 1987-01-19 12:00:00 1987-01-19 12:00:00
## 20  1987-01-20 12:00:00 1987-01-20 12:00:00 1987-01-20 12:00:00
## 21  1987-01-21 12:00:00 1987-01-21 12:00:00 1987-01-21 12:00:00
## 22  1987-01-22 12:00:00 1987-01-22 12:00:00 1987-01-22 12:00:00
## 23  1987-01-23 12:00:00 1987-01-23 12:00:00 1987-01-23 12:00:00
## 24  1987-01-24 12:00:00 1987-01-24 12:00:00 1987-01-24 12:00:00
## 25  1987-01-25 12:00:00 1987-01-25 12:00:00 1987-01-25 12:00:00
## 26  1987-01-26 12:00:00 1987-01-26 12:00:00 1987-01-26 12:00:00
## 27  1987-01-27 12:00:00 1987-01-27 12:00:00 1987-01-27 12:00:00
## 28  1987-01-28 12:00:00 1987-01-28 12:00:00 1987-01-28 12:00:00
## 29  1987-01-29 12:00:00 1987-01-29 12:00:00 1987-01-29 12:00:00
## 30  1987-01-30 12:00:00 1987-01-30 12:00:00 1987-01-30 12:00:00
## 31  1987-01-31 12:00:00 1987-01-31 12:00:00 1987-01-31 12:00:00
## 32  1987-02-01 12:00:00 1987-02-01 12:00:00 1987-02-01 12:00:00
## 33  1987-02-02 12:00:00 1987-02-02 12:00:00 1987-02-02 12:00:00
## 34  1987-02-03 12:00:00 1987-02-03 12:00:00 1987-02-03 12:00:00
## 35  1987-02-04 12:00:00 1987-02-04 12:00:00 1987-02-04 12:00:00
## 36  1987-02-05 12:00:00 1987-02-05 12:00:00 1987-02-05 12:00:00
## 37  1987-02-06 12:00:00 1987-02-06 12:00:00 1987-02-06 12:00:00
## 38  1987-02-07 12:00:00 1987-02-07 12:00:00 1987-02-07 12:00:00
## 39  1987-02-08 12:00:00 1987-02-08 12:00:00 1987-02-08 12:00:00
## 40  1987-02-09 12:00:00 1987-02-09 12:00:00 1987-02-09 12:00:00
## 41  1987-02-10 12:00:00 1987-02-10 12:00:00 1987-02-10 12:00:00
## 42  1987-02-11 12:00:00 1987-02-11 12:00:00 1987-02-11 12:00:00
## 43  1987-02-12 12:00:00 1987-02-12 12:00:00 1987-02-12 12:00:00
## 44  1987-02-13 12:00:00 1987-02-13 12:00:00 1987-02-13 12:00:00
## 45  1987-02-14 12:00:00 1987-02-14 12:00:00 1987-02-14 12:00:00
## 46  1987-02-15 12:00:00 1987-02-15 12:00:00 1987-02-15 12:00:00
## 47  1987-02-16 12:00:00 1987-02-16 12:00:00 1987-02-16 12:00:00
## 48  1987-02-17 12:00:00 1987-02-17 12:00:00 1987-02-17 12:00:00
## 49  1987-02-18 12:00:00 1987-02-18 12:00:00 1987-02-18 12:00:00
## 50  1987-02-19 12:00:00 1987-02-19 12:00:00 1987-02-19 12:00:00
## 51  1987-02-20 12:00:00 1987-02-20 12:00:00 1987-02-20 12:00:00
## 52  1987-02-21 12:00:00 1987-02-21 12:00:00 1987-02-21 12:00:00
## 53  1987-02-22 12:00:00 1987-02-22 12:00:00 1987-02-22 12:00:00
## 54  1987-02-23 12:00:00 1987-02-23 12:00:00 1987-02-23 12:00:00
## 55  1987-02-24 12:00:00 1987-02-24 12:00:00 1987-02-24 12:00:00
## 56  1987-02-25 12:00:00 1987-02-25 12:00:00 1987-02-25 12:00:00
## 57  1987-02-26 12:00:00 1987-02-26 12:00:00 1987-02-26 12:00:00
## 58  1987-02-27 12:00:00 1987-02-27 12:00:00 1987-02-27 12:00:00
## 59  1987-02-28 12:00:00 1987-02-28 12:00:00 1987-02-28 12:00:00
## 60  1987-03-01 12:00:00 1987-03-01 12:00:00 1987-03-01 12:00:00
## 61  1987-03-02 12:00:00 1987-03-02 12:00:00 1987-03-02 12:00:00
## 62  1987-03-03 12:00:00 1987-03-03 12:00:00 1987-03-03 12:00:00
## 63  1987-03-04 12:00:00 1987-03-04 12:00:00 1987-03-04 12:00:00
## 64  1987-03-05 12:00:00 1987-03-05 12:00:00 1987-03-05 12:00:00
## 65  1987-03-06 12:00:00 1987-03-06 12:00:00 1987-03-06 12:00:00
## 66  1987-03-07 12:00:00 1987-03-07 12:00:00 1987-03-07 12:00:00
## 67  1987-03-08 12:00:00 1987-03-08 12:00:00 1987-03-08 12:00:00
## 68  1987-03-09 12:00:00 1987-03-09 12:00:00 1987-03-09 12:00:00
## 69  1987-03-10 12:00:00 1987-03-10 12:00:00 1987-03-10 12:00:00
## 70  1987-03-11 12:00:00 1987-03-11 12:00:00 1987-03-11 12:00:00
## 71  1987-03-12 12:00:00 1987-03-12 12:00:00 1987-03-12 12:00:00
## 72  1987-03-13 12:00:00 1987-03-13 12:00:00 1987-03-13 12:00:00
## 73  1987-03-14 12:00:00 1987-03-14 12:00:00 1987-03-14 12:00:00
## 74  1987-03-15 12:00:00 1987-03-15 12:00:00 1987-03-15 12:00:00
## 75  1987-03-16 12:00:00 1987-03-16 12:00:00 1987-03-16 12:00:00
## 76  1987-03-17 12:00:00 1987-03-17 12:00:00 1987-03-17 12:00:00
## 77  1987-03-18 12:00:00 1987-03-18 12:00:00 1987-03-18 12:00:00
## 78  1987-03-19 12:00:00 1987-03-19 12:00:00 1987-03-19 12:00:00
## 79  1987-03-20 12:00:00 1987-03-20 12:00:00 1987-03-20 12:00:00
## 80  1987-03-21 12:00:00 1987-03-21 12:00:00 1987-03-21 12:00:00
## 81  1987-03-22 12:00:00 1987-03-22 12:00:00 1987-03-22 12:00:00
## 82  1987-03-23 12:00:00 1987-03-23 12:00:00 1987-03-23 12:00:00
## 83  1987-03-24 12:00:00 1987-03-24 12:00:00 1987-03-24 12:00:00
## 84  1987-03-25 12:00:00 1987-03-25 12:00:00 1987-03-25 12:00:00
## 85  1987-03-26 12:00:00 1987-03-26 12:00:00 1987-03-26 12:00:00
## 86  1987-03-27 12:00:00 1987-03-27 12:00:00 1987-03-27 12:00:00
## 87  1987-03-28 12:00:00 1987-03-28 12:00:00 1987-03-28 12:00:00
## 88  1987-03-29 12:00:00 1987-03-29 12:00:00 1987-03-29 12:00:00
## 89  1987-03-30 12:00:00 1987-03-30 12:00:00 1987-03-30 12:00:00
## 90  1987-03-31 12:00:00 1987-03-31 12:00:00 1987-03-31 12:00:00
## 91  1987-04-01 12:00:00 1987-04-01 12:00:00 1987-04-01 12:00:00
## 92  1987-04-02 12:00:00 1987-04-02 12:00:00 1987-04-02 12:00:00
## 93  1987-04-03 12:00:00 1987-04-03 12:00:00 1987-04-03 12:00:00
## 94  1987-04-04 12:00:00 1987-04-04 12:00:00 1987-04-04 12:00:00
## 95  1987-04-05 12:00:00 1987-04-05 12:00:00 1987-04-05 12:00:00
## 96  1987-04-06 12:00:00 1987-04-06 12:00:00 1987-04-06 12:00:00
## 97  1987-04-07 12:00:00 1987-04-07 12:00:00 1987-04-07 12:00:00
## 98  1987-04-08 12:00:00 1987-04-08 12:00:00 1987-04-08 12:00:00
## 99  1987-04-09 12:00:00 1987-04-09 12:00:00 1987-04-09 12:00:00
## 100 1987-04-10 12:00:00 1987-04-10 12:00:00 1987-04-10 12:00:00
## 101 1987-04-11 12:00:00 1987-04-11 12:00:00 1987-04-11 12:00:00
## 102 1987-04-12 12:00:00 1987-04-12 12:00:00 1987-04-12 12:00:00
## 103 1987-04-13 12:00:00 1987-04-13 12:00:00 1987-04-13 12:00:00
## 104 1987-04-14 12:00:00 1987-04-14 12:00:00 1987-04-14 12:00:00
## 105 1987-04-15 12:00:00 1987-04-15 12:00:00 1987-04-15 12:00:00
## 106 1987-04-16 12:00:00 1987-04-16 12:00:00 1987-04-16 12:00:00
## 107 1987-04-17 12:00:00 1987-04-17 12:00:00 1987-04-17 12:00:00
## 108 1987-04-18 12:00:00 1987-04-18 12:00:00 1987-04-18 12:00:00
## 109 1987-04-19 12:00:00 1987-04-19 12:00:00 1987-04-19 12:00:00
## 110 1987-04-20 12:00:00 1987-04-20 12:00:00 1987-04-20 12:00:00
## 111 1987-04-21 12:00:00 1987-04-21 12:00:00 1987-04-21 12:00:00
## 112 1987-04-22 12:00:00 1987-04-22 12:00:00 1987-04-22 12:00:00
## 113 1987-04-23 12:00:00 1987-04-23 12:00:00 1987-04-23 12:00:00
## 114 1987-04-24 12:00:00 1987-04-24 12:00:00 1987-04-24 12:00:00
## 115 1987-04-25 12:00:00 1987-04-25 12:00:00 1987-04-25 12:00:00
## 116 1987-04-26 12:00:00 1987-04-26 12:00:00 1987-04-26 12:00:00
## 117 1987-04-27 12:00:00 1987-04-27 12:00:00 1987-04-27 12:00:00
## 118 1987-04-28 12:00:00 1987-04-28 12:00:00 1987-04-28 12:00:00
## 119 1987-04-29 12:00:00 1987-04-29 12:00:00 1987-04-29 12:00:00
## 120 1987-04-30 12:00:00 1987-04-30 12:00:00 1987-04-30 12:00:00
## 121 1987-05-01 12:00:00 1987-05-01 12:00:00 1987-05-01 12:00:00
## 122 1987-05-02 12:00:00 1987-05-02 12:00:00 1987-05-02 12:00:00
## 123 1987-05-03 12:00:00 1987-05-03 12:00:00 1987-05-03 12:00:00
## 124 1987-05-04 12:00:00 1987-05-04 12:00:00 1987-05-04 12:00:00
## 125 1987-05-05 12:00:00 1987-05-05 12:00:00 1987-05-05 12:00:00
## 126 1987-05-06 12:00:00 1987-05-06 12:00:00 1987-05-06 12:00:00
## 127 1987-05-07 12:00:00 1987-05-07 12:00:00 1987-05-07 12:00:00
## 128 1987-05-08 12:00:00 1987-05-08 12:00:00 1987-05-08 12:00:00
## 129 1987-05-09 12:00:00 1987-05-09 12:00:00 1987-05-09 12:00:00
## 130 1987-05-10 12:00:00 1987-05-10 12:00:00 1987-05-10 12:00:00
## 131 1987-05-11 12:00:00 1987-05-11 12:00:00 1987-05-11 12:00:00
## 132 1987-05-12 12:00:00 1987-05-12 12:00:00 1987-05-12 12:00:00
## 133 1987-05-13 12:00:00 1987-05-13 12:00:00 1987-05-13 12:00:00
## 134 1987-05-14 12:00:00 1987-05-14 12:00:00 1987-05-14 12:00:00
## 135 1987-05-15 12:00:00 1987-05-15 12:00:00 1987-05-15 12:00:00
## 136 1987-05-16 12:00:00 1987-05-16 12:00:00 1987-05-16 12:00:00
## 137 1987-05-17 12:00:00 1987-05-17 12:00:00 1987-05-17 12:00:00
## 138 1987-05-18 12:00:00 1987-05-18 12:00:00 1987-05-18 12:00:00
## 139 1987-05-19 12:00:00 1987-05-19 12:00:00 1987-05-19 12:00:00
## 140 1987-05-20 12:00:00 1987-05-20 12:00:00 1987-05-20 12:00:00
## 141 1987-05-21 12:00:00 1987-05-21 12:00:00 1987-05-21 12:00:00
## 142 1987-05-22 12:00:00 1987-05-22 12:00:00 1987-05-22 12:00:00
## 143 1987-05-23 12:00:00 1987-05-23 12:00:00 1987-05-23 12:00:00
## 144 1987-05-24 12:00:00 1987-05-24 12:00:00 1987-05-24 12:00:00
## 145 1987-05-25 12:00:00 1987-05-25 12:00:00 1987-05-25 12:00:00
## 146 1987-05-26 12:00:00 1987-05-26 12:00:00 1987-05-26 12:00:00
## 147 1987-05-27 12:00:00 1987-05-27 12:00:00 1987-05-27 12:00:00
## 148 1987-05-28 12:00:00 1987-05-28 12:00:00 1987-05-28 12:00:00
## 149 1987-05-29 12:00:00 1987-05-29 12:00:00 1987-05-29 12:00:00
## 150 1987-05-30 12:00:00 1987-05-30 12:00:00 1987-05-30 12:00:00
## 151 1987-05-31 12:00:00 1987-05-31 12:00:00 1987-05-31 12:00:00
## 152 1987-06-01 12:00:00 1987-06-01 12:00:00 1987-06-01 12:00:00
## 153 1987-06-02 12:00:00 1987-06-02 12:00:00 1987-06-02 12:00:00
## 154 1987-06-03 12:00:00 1987-06-03 12:00:00 1987-06-03 12:00:00
## 155 1987-06-04 12:00:00 1987-06-04 12:00:00 1987-06-04 12:00:00
## 156 1987-06-05 12:00:00 1987-06-05 12:00:00 1987-06-05 12:00:00
## 157 1987-06-06 12:00:00 1987-06-06 12:00:00 1987-06-06 12:00:00
## 158 1987-06-07 12:00:00 1987-06-07 12:00:00 1987-06-07 12:00:00
## 159 1987-06-08 12:00:00 1987-06-08 12:00:00 1987-06-08 12:00:00
## 160 1987-06-09 12:00:00 1987-06-09 12:00:00 1987-06-09 12:00:00
## 161 1987-06-10 12:00:00 1987-06-10 12:00:00 1987-06-10 12:00:00
## 162 1987-06-11 12:00:00 1987-06-11 12:00:00 1987-06-11 12:00:00
## 163 1987-06-12 12:00:00 1987-06-12 12:00:00 1987-06-12 12:00:00
## 164 1987-06-13 12:00:00 1987-06-13 12:00:00 1987-06-13 12:00:00
## 165 1987-06-14 12:00:00 1987-06-14 12:00:00 1987-06-14 12:00:00
## 166 1987-06-15 12:00:00 1987-06-15 12:00:00 1987-06-15 12:00:00
## 167 1987-06-16 12:00:00 1987-06-16 12:00:00 1987-06-16 12:00:00
## 168 1987-06-17 12:00:00 1987-06-17 12:00:00 1987-06-17 12:00:00
## 169 1987-06-18 12:00:00 1987-06-18 12:00:00 1987-06-18 12:00:00
## 170 1987-06-19 12:00:00 1987-06-19 12:00:00 1987-06-19 12:00:00
## 171 1987-06-20 12:00:00 1987-06-20 12:00:00 1987-06-20 12:00:00
## 172 1987-06-21 12:00:00 1987-06-21 12:00:00 1987-06-21 12:00:00
## 173 1987-06-22 12:00:00 1987-06-22 12:00:00 1987-06-22 12:00:00
## 174 1987-06-23 12:00:00 1987-06-23 12:00:00 1987-06-23 12:00:00
## 175 1987-06-24 12:00:00 1987-06-24 12:00:00 1987-06-24 12:00:00
## 176 1987-06-25 12:00:00 1987-06-25 12:00:00 1987-06-25 12:00:00
## 177 1987-06-26 12:00:00 1987-06-26 12:00:00 1987-06-26 12:00:00
## 178 1987-06-27 12:00:00 1987-06-27 12:00:00 1987-06-27 12:00:00
## 179 1987-06-28 12:00:00 1987-06-28 12:00:00 1987-06-28 12:00:00
## 180 1987-06-29 12:00:00 1987-06-29 12:00:00 1987-06-29 12:00:00
## 181 1987-06-30 12:00:00 1987-06-30 12:00:00 1987-06-30 12:00:00
## 182 1987-07-01 12:00:00 1987-07-01 12:00:00 1987-07-01 12:00:00
## 183 1987-07-02 12:00:00 1987-07-02 12:00:00 1987-07-02 12:00:00
## 184 1987-07-03 12:00:00 1987-07-03 12:00:00 1987-07-03 12:00:00
## 185 1987-07-04 12:00:00 1987-07-04 12:00:00 1987-07-04 12:00:00
## 186 1987-07-05 12:00:00 1987-07-05 12:00:00 1987-07-05 12:00:00
## 187 1987-07-06 12:00:00 1987-07-06 12:00:00 1987-07-06 12:00:00
## 188 1987-07-07 12:00:00 1987-07-07 12:00:00 1987-07-07 12:00:00
## 189 1987-07-08 12:00:00 1987-07-08 12:00:00 1987-07-08 12:00:00
## 190 1987-07-09 12:00:00 1987-07-09 12:00:00 1987-07-09 12:00:00
## 191 1987-07-10 12:00:00 1987-07-10 12:00:00 1987-07-10 12:00:00
## 192 1987-07-11 12:00:00 1987-07-11 12:00:00 1987-07-11 12:00:00
## 193 1987-07-12 12:00:00 1987-07-12 12:00:00 1987-07-12 12:00:00
## 194 1987-07-13 12:00:00 1987-07-13 12:00:00 1987-07-13 12:00:00
## 195 1987-07-14 12:00:00 1987-07-14 12:00:00 1987-07-14 12:00:00
## 196 1987-07-15 12:00:00 1987-07-15 12:00:00 1987-07-15 12:00:00
## 197 1987-07-16 12:00:00 1987-07-16 12:00:00 1987-07-16 12:00:00
## 198 1987-07-17 12:00:00 1987-07-17 12:00:00 1987-07-17 12:00:00
## 199 1987-07-18 12:00:00 1987-07-18 12:00:00 1987-07-18 12:00:00
## 200 1987-07-19 12:00:00 1987-07-19 12:00:00 1987-07-19 12:00:00
## 201 1987-07-20 12:00:00 1987-07-20 12:00:00 1987-07-20 12:00:00
## 202 1987-07-21 12:00:00 1987-07-21 12:00:00 1987-07-21 12:00:00
## 203 1987-07-22 12:00:00 1987-07-22 12:00:00 1987-07-22 12:00:00
## 204 1987-07-23 12:00:00 1987-07-23 12:00:00 1987-07-23 12:00:00
## 205 1987-07-24 12:00:00 1987-07-24 12:00:00 1987-07-24 12:00:00
## 206 1987-07-25 12:00:00 1987-07-25 12:00:00 1987-07-25 12:00:00
## 207 1987-07-26 12:00:00 1987-07-26 12:00:00 1987-07-26 12:00:00
## 208 1987-07-27 12:00:00 1987-07-27 12:00:00 1987-07-27 12:00:00
## 209 1987-07-28 12:00:00 1987-07-28 12:00:00 1987-07-28 12:00:00
## 210 1987-07-29 12:00:00 1987-07-29 12:00:00 1987-07-29 12:00:00
## 211 1987-07-30 12:00:00 1987-07-30 12:00:00 1987-07-30 12:00:00
## 212 1987-07-31 12:00:00 1987-07-31 12:00:00 1987-07-31 12:00:00
## 213 1987-08-01 12:00:00 1987-08-01 12:00:00 1987-08-01 12:00:00
## 214 1987-08-02 12:00:00 1987-08-02 12:00:00 1987-08-02 12:00:00
## 215 1987-08-03 12:00:00 1987-08-03 12:00:00 1987-08-03 12:00:00
## 216 1987-08-04 12:00:00 1987-08-04 12:00:00 1987-08-04 12:00:00
## 217 1987-08-05 12:00:00 1987-08-05 12:00:00 1987-08-05 12:00:00
## 218 1987-08-06 12:00:00 1987-08-06 12:00:00 1987-08-06 12:00:00
## 219 1987-08-07 12:00:00 1987-08-07 12:00:00 1987-08-07 12:00:00
## 220 1987-08-08 12:00:00 1987-08-08 12:00:00 1987-08-08 12:00:00
## 221 1987-08-09 12:00:00 1987-08-09 12:00:00 1987-08-09 12:00:00
## 222 1987-08-10 12:00:00 1987-08-10 12:00:00 1987-08-10 12:00:00
## 223 1987-08-11 12:00:00 1987-08-11 12:00:00 1987-08-11 12:00:00
## 224 1987-08-12 12:00:00 1987-08-12 12:00:00 1987-08-12 12:00:00
## 225 1987-08-13 12:00:00 1987-08-13 12:00:00 1987-08-13 12:00:00
## 226 1987-08-14 12:00:00 1987-08-14 12:00:00 1987-08-14 12:00:00
## 227 1987-08-15 12:00:00 1987-08-15 12:00:00 1987-08-15 12:00:00
## 228 1987-08-16 12:00:00 1987-08-16 12:00:00 1987-08-16 12:00:00
## 229 1987-08-17 12:00:00 1987-08-17 12:00:00 1987-08-17 12:00:00
## 230 1987-08-18 12:00:00 1987-08-18 12:00:00 1987-08-18 12:00:00
## 231 1987-08-19 12:00:00 1987-08-19 12:00:00 1987-08-19 12:00:00
## 232 1987-08-20 12:00:00 1987-08-20 12:00:00 1987-08-20 12:00:00
## 233 1987-08-21 12:00:00 1987-08-21 12:00:00 1987-08-21 12:00:00
## 234 1987-08-22 12:00:00 1987-08-22 12:00:00 1987-08-22 12:00:00
## 235 1987-08-23 12:00:00 1987-08-23 12:00:00 1987-08-23 12:00:00
## 236 1987-08-24 12:00:00 1987-08-24 12:00:00 1987-08-24 12:00:00
## 237 1987-08-25 12:00:00 1987-08-25 12:00:00 1987-08-25 12:00:00
## 238 1987-08-26 12:00:00 1987-08-26 12:00:00 1987-08-26 12:00:00
## 239 1987-08-27 12:00:00 1987-08-27 12:00:00 1987-08-27 12:00:00
## 240 1987-08-28 12:00:00 1987-08-28 12:00:00 1987-08-28 12:00:00
## 241 1987-08-29 12:00:00 1987-08-29 12:00:00 1987-08-29 12:00:00
## 242 1987-08-30 12:00:00 1987-08-30 12:00:00 1987-08-30 12:00:00
## 243 1987-08-31 12:00:00 1987-08-31 12:00:00 1987-08-31 12:00:00
## 244 1987-09-01 12:00:00 1987-09-01 12:00:00 1987-09-01 12:00:00
## 245 1987-09-02 12:00:00 1987-09-02 12:00:00 1987-09-02 12:00:00
## 246 1987-09-03 12:00:00 1987-09-03 12:00:00 1987-09-03 12:00:00
## 247 1987-09-04 12:00:00 1987-09-04 12:00:00 1987-09-04 12:00:00
## 248 1987-09-05 12:00:00 1987-09-05 12:00:00 1987-09-05 12:00:00
## 249 1987-09-06 12:00:00 1987-09-06 12:00:00 1987-09-06 12:00:00
## 250 1987-09-07 12:00:00 1987-09-07 12:00:00 1987-09-07 12:00:00
## 251 1987-09-08 12:00:00 1987-09-08 12:00:00 1987-09-08 12:00:00
## 252 1987-09-09 12:00:00 1987-09-09 12:00:00 1987-09-09 12:00:00
## 253 1987-09-10 12:00:00 1987-09-10 12:00:00 1987-09-10 12:00:00
## 254 1987-09-11 12:00:00 1987-09-11 12:00:00 1987-09-11 12:00:00
## 255 1987-09-12 12:00:00 1987-09-12 12:00:00 1987-09-12 12:00:00
## 256 1987-09-13 12:00:00 1987-09-13 12:00:00 1987-09-13 12:00:00
## 257 1987-09-14 12:00:00 1987-09-14 12:00:00 1987-09-14 12:00:00
## 258 1987-09-15 12:00:00 1987-09-15 12:00:00 1987-09-15 12:00:00
## 259 1987-09-16 12:00:00 1987-09-16 12:00:00 1987-09-16 12:00:00
## 260 1987-09-17 12:00:00 1987-09-17 12:00:00 1987-09-17 12:00:00
## 261 1987-09-18 12:00:00 1987-09-18 12:00:00 1987-09-18 12:00:00
## 262 1987-09-19 12:00:00 1987-09-19 12:00:00 1987-09-19 12:00:00
## 263 1987-09-20 12:00:00 1987-09-20 12:00:00 1987-09-20 12:00:00
## 264 1987-09-21 12:00:00 1987-09-21 12:00:00 1987-09-21 12:00:00
## 265 1987-09-22 12:00:00 1987-09-22 12:00:00 1987-09-22 12:00:00
## 266 1987-09-23 12:00:00 1987-09-23 12:00:00 1987-09-23 12:00:00
## 267 1987-09-24 12:00:00 1987-09-24 12:00:00 1987-09-24 12:00:00
## 268 1987-09-25 12:00:00 1987-09-25 12:00:00 1987-09-25 12:00:00
## 269 1987-09-26 12:00:00 1987-09-26 12:00:00 1987-09-26 12:00:00
## 270 1987-09-27 12:00:00 1987-09-27 12:00:00 1987-09-27 12:00:00
## 271 1987-09-28 12:00:00 1987-09-28 12:00:00 1987-09-28 12:00:00
## 272 1987-09-29 12:00:00 1987-09-29 12:00:00 1987-09-29 12:00:00
## 273 1987-09-30 12:00:00 1987-09-30 12:00:00 1987-09-30 12:00:00
## 274 1987-10-01 12:00:00 1987-10-01 12:00:00 1987-10-01 12:00:00
## 275 1987-10-02 12:00:00 1987-10-02 12:00:00 1987-10-02 12:00:00
## 276 1987-10-03 12:00:00 1987-10-03 12:00:00 1987-10-03 12:00:00
## 277 1987-10-04 12:00:00 1987-10-04 12:00:00 1987-10-04 12:00:00
## 278 1987-10-05 12:00:00 1987-10-05 12:00:00 1987-10-05 12:00:00
## 279 1987-10-06 12:00:00 1987-10-06 12:00:00 1987-10-06 12:00:00
## 280 1987-10-07 12:00:00 1987-10-07 12:00:00 1987-10-07 12:00:00
## 281 1987-10-08 12:00:00 1987-10-08 12:00:00 1987-10-08 12:00:00
## 282 1987-10-09 12:00:00 1987-10-09 12:00:00 1987-10-09 12:00:00
## 283 1987-10-10 12:00:00 1987-10-10 12:00:00 1987-10-10 12:00:00
## 284 1987-10-11 12:00:00 1987-10-11 12:00:00 1987-10-11 12:00:00
## 285 1987-10-12 12:00:00 1987-10-12 12:00:00 1987-10-12 12:00:00
## 286 1987-10-13 12:00:00 1987-10-13 12:00:00 1987-10-13 12:00:00
## 287 1987-10-14 12:00:00 1987-10-14 12:00:00 1987-10-14 12:00:00
## 288 1987-10-15 12:00:00 1987-10-15 12:00:00 1987-10-15 12:00:00
## 289 1987-10-16 12:00:00 1987-10-16 12:00:00 1987-10-16 12:00:00
## 290 1987-10-17 12:00:00 1987-10-17 12:00:00 1987-10-17 12:00:00
## 291 1987-10-18 12:00:00 1987-10-18 12:00:00 1987-10-18 12:00:00
## 292 1987-10-19 12:00:00 1987-10-19 12:00:00 1987-10-19 12:00:00
## 293 1987-10-20 12:00:00 1987-10-20 12:00:00 1987-10-20 12:00:00
## 294 1987-10-21 12:00:00 1987-10-21 12:00:00 1987-10-21 12:00:00
## 295 1987-10-22 12:00:00 1987-10-22 12:00:00 1987-10-22 12:00:00
## 296 1987-10-23 12:00:00 1987-10-23 12:00:00 1987-10-23 12:00:00
## 297 1987-10-24 12:00:00 1987-10-24 12:00:00 1987-10-24 12:00:00
## 298 1987-10-25 12:00:00 1987-10-25 12:00:00 1987-10-25 12:00:00
## 299 1987-10-26 12:00:00 1987-10-26 12:00:00 1987-10-26 12:00:00
## 300 1987-10-27 12:00:00 1987-10-27 12:00:00 1987-10-27 12:00:00
## 301 1987-10-28 12:00:00 1987-10-28 12:00:00 1987-10-28 12:00:00
## 302 1987-10-29 12:00:00 1987-10-29 12:00:00 1987-10-29 12:00:00
## 303 1987-10-30 12:00:00 1987-10-30 12:00:00 1987-10-30 12:00:00
## 304 1987-10-31 12:00:00 1987-10-31 12:00:00 1987-10-31 12:00:00
## 305 1987-11-01 12:00:00 1987-11-01 12:00:00 1987-11-01 12:00:00
## 306 1987-11-02 12:00:00 1987-11-02 12:00:00 1987-11-02 12:00:00
## 307 1987-11-03 12:00:00 1987-11-03 12:00:00 1987-11-03 12:00:00
## 308 1987-11-04 12:00:00 1987-11-04 12:00:00 1987-11-04 12:00:00
## 309 1987-11-05 12:00:00 1987-11-05 12:00:00 1987-11-05 12:00:00
## 310 1987-11-06 12:00:00 1987-11-06 12:00:00 1987-11-06 12:00:00
## 311 1987-11-07 12:00:00 1987-11-07 12:00:00 1987-11-07 12:00:00
## 312 1987-11-08 12:00:00 1987-11-08 12:00:00 1987-11-08 12:00:00
## 313 1987-11-09 12:00:00 1987-11-09 12:00:00 1987-11-09 12:00:00
## 314 1987-11-10 12:00:00 1987-11-10 12:00:00 1987-11-10 12:00:00
## 315 1987-11-11 12:00:00 1987-11-11 12:00:00 1987-11-11 12:00:00
## 316 1987-11-12 12:00:00 1987-11-12 12:00:00 1987-11-12 12:00:00
## 317 1987-11-13 12:00:00 1987-11-13 12:00:00 1987-11-13 12:00:00
## 318 1987-11-14 12:00:00 1987-11-14 12:00:00 1987-11-14 12:00:00
## 319 1987-11-15 12:00:00 1987-11-15 12:00:00 1987-11-15 12:00:00
## 320 1987-11-16 12:00:00 1987-11-16 12:00:00 1987-11-16 12:00:00
## 321 1987-11-17 12:00:00 1987-11-17 12:00:00 1987-11-17 12:00:00
## 322 1987-11-18 12:00:00 1987-11-18 12:00:00 1987-11-18 12:00:00
## 323 1987-11-19 12:00:00 1987-11-19 12:00:00 1987-11-19 12:00:00
## 324 1987-11-20 12:00:00 1987-11-20 12:00:00 1987-11-20 12:00:00
## 325 1987-11-21 12:00:00 1987-11-21 12:00:00 1987-11-21 12:00:00
## 326 1987-11-22 12:00:00 1987-11-22 12:00:00 1987-11-22 12:00:00
## 327 1987-11-23 12:00:00 1987-11-23 12:00:00 1987-11-23 12:00:00
## 328 1987-11-24 12:00:00 1987-11-24 12:00:00 1987-11-24 12:00:00
## 329 1987-11-25 12:00:00 1987-11-25 12:00:00 1987-11-25 12:00:00
## 330 1987-11-26 12:00:00 1987-11-26 12:00:00 1987-11-26 12:00:00
## 331 1987-11-27 12:00:00 1987-11-27 12:00:00 1987-11-27 12:00:00
## 332 1987-11-28 12:00:00 1987-11-28 12:00:00 1987-11-28 12:00:00
## 333 1987-11-29 12:00:00 1987-11-29 12:00:00 1987-11-29 12:00:00
## 334 1987-11-30 12:00:00 1987-11-30 12:00:00 1987-11-30 12:00:00
## 335 1987-12-01 12:00:00 1987-12-01 12:00:00 1987-12-01 12:00:00
## 336 1987-12-02 12:00:00 1987-12-02 12:00:00 1987-12-02 12:00:00
## 337 1987-12-03 12:00:00 1987-12-03 12:00:00 1987-12-03 12:00:00
## 338 1987-12-04 12:00:00 1987-12-04 12:00:00 1987-12-04 12:00:00
## 339 1987-12-05 12:00:00 1987-12-05 12:00:00 1987-12-05 12:00:00
## 340 1987-12-06 12:00:00 1987-12-06 12:00:00 1987-12-06 12:00:00
## 341 1987-12-07 12:00:00 1987-12-07 12:00:00 1987-12-07 12:00:00
## 342 1987-12-08 12:00:00 1987-12-08 12:00:00 1987-12-08 12:00:00
## 343 1987-12-09 12:00:00 1987-12-09 12:00:00 1987-12-09 12:00:00
## 344 1987-12-10 12:00:00 1987-12-10 12:00:00 1987-12-10 12:00:00
## 345 1987-12-11 12:00:00 1987-12-11 12:00:00 1987-12-11 12:00:00
## 346 1987-12-12 12:00:00 1987-12-12 12:00:00 1987-12-12 12:00:00
## 347 1987-12-13 12:00:00 1987-12-13 12:00:00 1987-12-13 12:00:00
## 348 1987-12-14 12:00:00 1987-12-14 12:00:00 1987-12-14 12:00:00
## 349 1987-12-15 12:00:00 1987-12-15 12:00:00 1987-12-15 12:00:00
## 350 1987-12-16 12:00:00 1987-12-16 12:00:00 1987-12-16 12:00:00
## 351 1987-12-17 12:00:00 1987-12-17 12:00:00 1987-12-17 12:00:00
## 352 1987-12-18 12:00:00 1987-12-18 12:00:00 1987-12-18 12:00:00
## 353 1987-12-19 12:00:00 1987-12-19 12:00:00 1987-12-19 12:00:00
## 354 1987-12-20 12:00:00 1987-12-20 12:00:00 1987-12-20 12:00:00
## 355 1987-12-21 12:00:00 1987-12-21 12:00:00 1987-12-21 12:00:00
## 356 1987-12-22 12:00:00 1987-12-22 12:00:00 1987-12-22 12:00:00
## 357 1987-12-23 12:00:00 1987-12-23 12:00:00 1987-12-23 12:00:00
## 358 1987-12-24 12:00:00 1987-12-24 12:00:00 1987-12-24 12:00:00
## 359 1987-12-25 12:00:00 1987-12-25 12:00:00 1987-12-25 12:00:00
## 360 1987-12-26 12:00:00 1987-12-26 12:00:00 1987-12-26 12:00:00
## 361 1987-12-27 12:00:00 1987-12-27 12:00:00 1987-12-27 12:00:00
## 362 1987-12-28 12:00:00 1987-12-28 12:00:00 1987-12-28 12:00:00
## 363 1987-12-29 12:00:00 1987-12-29 12:00:00 1987-12-29 12:00:00
## 364 1987-12-30 12:00:00 1987-12-30 12:00:00 1987-12-30 12:00:00
## 365 1987-12-31 12:00:00 1987-12-31 12:00:00 1987-12-31 12:00:00
## 366 1988-01-01 12:00:00 1988-01-01 12:00:00 1988-01-01 12:00:00
## 367 1988-01-02 12:00:00 1988-01-02 12:00:00 1988-01-02 12:00:00
## 368 1988-01-03 12:00:00 1988-01-03 12:00:00 1988-01-03 12:00:00
## 369 1988-01-04 12:00:00 1988-01-04 12:00:00 1988-01-04 12:00:00
## 370 1988-01-05 12:00:00 1988-01-05 12:00:00 1988-01-05 12:00:00
## 371 1988-01-06 12:00:00 1988-01-06 12:00:00 1988-01-06 12:00:00
## 372 1988-01-07 12:00:00 1988-01-07 12:00:00 1988-01-07 12:00:00
## 373 1988-01-08 12:00:00 1988-01-08 12:00:00 1988-01-08 12:00:00
## 374 1988-01-09 12:00:00 1988-01-09 12:00:00 1988-01-09 12:00:00
## 375 1988-01-10 12:00:00 1988-01-10 12:00:00 1988-01-10 12:00:00
## 376 1988-01-11 12:00:00 1988-01-11 12:00:00 1988-01-11 12:00:00
## 377 1988-01-12 12:00:00 1988-01-12 12:00:00 1988-01-12 12:00:00
## 378 1988-01-13 12:00:00 1988-01-13 12:00:00 1988-01-13 12:00:00
## 379 1988-01-14 12:00:00 1988-01-14 12:00:00 1988-01-14 12:00:00
## 380 1988-01-15 12:00:00 1988-01-15 12:00:00 1988-01-15 12:00:00
## 381 1988-01-16 12:00:00 1988-01-16 12:00:00 1988-01-16 12:00:00
## 382 1988-01-17 12:00:00 1988-01-17 12:00:00 1988-01-17 12:00:00
## 383 1988-01-18 12:00:00 1988-01-18 12:00:00 1988-01-18 12:00:00
## 384 1988-01-19 12:00:00 1988-01-19 12:00:00 1988-01-19 12:00:00
## 385 1988-01-20 12:00:00 1988-01-20 12:00:00 1988-01-20 12:00:00
## 386 1988-01-21 12:00:00 1988-01-21 12:00:00 1988-01-21 12:00:00
## 387 1988-01-22 12:00:00 1988-01-22 12:00:00 1988-01-22 12:00:00
## 388 1988-01-23 12:00:00 1988-01-23 12:00:00 1988-01-23 12:00:00
## 389 1988-01-24 12:00:00 1988-01-24 12:00:00 1988-01-24 12:00:00
## 390 1988-01-25 12:00:00 1988-01-25 12:00:00 1988-01-25 12:00:00
## 391 1988-01-26 12:00:00 1988-01-26 12:00:00 1988-01-26 12:00:00
## 392 1988-01-27 12:00:00 1988-01-27 12:00:00 1988-01-27 12:00:00
## 393 1988-01-28 12:00:00 1988-01-28 12:00:00 1988-01-28 12:00:00
## 394 1988-01-29 12:00:00 1988-01-29 12:00:00 1988-01-29 12:00:00
## 395 1988-01-30 12:00:00 1988-01-30 12:00:00 1988-01-30 12:00:00
## 396 1988-01-31 12:00:00 1988-01-31 12:00:00 1988-01-31 12:00:00
## 397 1988-02-01 12:00:00 1988-02-01 12:00:00 1988-02-01 12:00:00
## 398 1988-02-02 12:00:00 1988-02-02 12:00:00 1988-02-02 12:00:00
## 399 1988-02-03 12:00:00 1988-02-03 12:00:00 1988-02-03 12:00:00
## 400 1988-02-04 12:00:00 1988-02-04 12:00:00 1988-02-04 12:00:00
## 401 1988-02-05 12:00:00 1988-02-05 12:00:00 1988-02-05 12:00:00
## 402 1988-02-06 12:00:00 1988-02-06 12:00:00 1988-02-06 12:00:00
## 403 1988-02-07 12:00:00 1988-02-07 12:00:00 1988-02-07 12:00:00
## 404 1988-02-08 12:00:00 1988-02-08 12:00:00 1988-02-08 12:00:00
## 405 1988-02-09 12:00:00 1988-02-09 12:00:00 1988-02-09 12:00:00
## 406 1988-02-10 12:00:00 1988-02-10 12:00:00 1988-02-10 12:00:00
## 407 1988-02-11 12:00:00 1988-02-11 12:00:00 1988-02-11 12:00:00
## 408 1988-02-12 12:00:00 1988-02-12 12:00:00 1988-02-12 12:00:00
## 409 1988-02-13 12:00:00 1988-02-13 12:00:00 1988-02-13 12:00:00
## 410 1988-02-14 12:00:00 1988-02-14 12:00:00 1988-02-14 12:00:00
## 411 1988-02-15 12:00:00 1988-02-15 12:00:00 1988-02-15 12:00:00
## 412 1988-02-16 12:00:00 1988-02-16 12:00:00 1988-02-16 12:00:00
## 413 1988-02-17 12:00:00 1988-02-17 12:00:00 1988-02-17 12:00:00
## 414 1988-02-18 12:00:00 1988-02-18 12:00:00 1988-02-18 12:00:00
## 415 1988-02-19 12:00:00 1988-02-19 12:00:00 1988-02-19 12:00:00
## 416 1988-02-20 12:00:00 1988-02-20 12:00:00 1988-02-20 12:00:00
## 417 1988-02-21 12:00:00 1988-02-21 12:00:00 1988-02-21 12:00:00
## 418 1988-02-22 12:00:00 1988-02-22 12:00:00 1988-02-22 12:00:00
## 419 1988-02-23 12:00:00 1988-02-23 12:00:00 1988-02-23 12:00:00
## 420 1988-02-24 12:00:00 1988-02-24 12:00:00 1988-02-24 12:00:00
## 421 1988-02-25 12:00:00 1988-02-25 12:00:00 1988-02-25 12:00:00
## 422 1988-02-26 12:00:00 1988-02-26 12:00:00 1988-02-26 12:00:00
## 423 1988-02-27 12:00:00 1988-02-27 12:00:00 1988-02-27 12:00:00
## 424 1988-02-28 12:00:00 1988-02-28 12:00:00 1988-02-28 12:00:00
## 425 1988-02-29 12:00:00 1988-02-29 12:00:00 1988-02-29 12:00:00
## 426 1988-03-01 12:00:00 1988-03-01 12:00:00 1988-03-01 12:00:00
## 427 1988-03-02 12:00:00 1988-03-02 12:00:00 1988-03-02 12:00:00
## 428 1988-03-03 12:00:00 1988-03-03 12:00:00 1988-03-03 12:00:00
## 429 1988-03-04 12:00:00 1988-03-04 12:00:00 1988-03-04 12:00:00
## 430 1988-03-05 12:00:00 1988-03-05 12:00:00 1988-03-05 12:00:00
## 431 1988-03-06 12:00:00 1988-03-06 12:00:00 1988-03-06 12:00:00
## 432 1988-03-07 12:00:00 1988-03-07 12:00:00 1988-03-07 12:00:00
## 433 1988-03-08 12:00:00 1988-03-08 12:00:00 1988-03-08 12:00:00
## 434 1988-03-09 12:00:00 1988-03-09 12:00:00 1988-03-09 12:00:00
## 435 1988-03-10 12:00:00 1988-03-10 12:00:00 1988-03-10 12:00:00
## 436 1988-03-11 12:00:00 1988-03-11 12:00:00 1988-03-11 12:00:00
## 437 1988-03-12 12:00:00 1988-03-12 12:00:00 1988-03-12 12:00:00
## 438 1988-03-13 12:00:00 1988-03-13 12:00:00 1988-03-13 12:00:00
## 439 1988-03-14 12:00:00 1988-03-14 12:00:00 1988-03-14 12:00:00
## 440 1988-03-15 12:00:00 1988-03-15 12:00:00 1988-03-15 12:00:00
## 441 1988-03-16 12:00:00 1988-03-16 12:00:00 1988-03-16 12:00:00
## 442 1988-03-17 12:00:00 1988-03-17 12:00:00 1988-03-17 12:00:00
## 443 1988-03-18 12:00:00 1988-03-18 12:00:00 1988-03-18 12:00:00
## 444 1988-03-19 12:00:00 1988-03-19 12:00:00 1988-03-19 12:00:00
## 445 1988-03-20 12:00:00 1988-03-20 12:00:00 1988-03-20 12:00:00
## 446 1988-03-21 12:00:00 1988-03-21 12:00:00 1988-03-21 12:00:00
## 447 1988-03-22 12:00:00 1988-03-22 12:00:00 1988-03-22 12:00:00
## 448 1988-03-23 12:00:00 1988-03-23 12:00:00 1988-03-23 12:00:00
## 449 1988-03-24 12:00:00 1988-03-24 12:00:00 1988-03-24 12:00:00
## 450 1988-03-25 12:00:00 1988-03-25 12:00:00 1988-03-25 12:00:00
## 451 1988-03-26 12:00:00 1988-03-26 12:00:00 1988-03-26 12:00:00
## 452 1988-03-27 12:00:00 1988-03-27 12:00:00 1988-03-27 12:00:00
## 453 1988-03-28 12:00:00 1988-03-28 12:00:00 1988-03-28 12:00:00
## 454 1988-03-29 12:00:00 1988-03-29 12:00:00 1988-03-29 12:00:00
## 455 1988-03-30 12:00:00 1988-03-30 12:00:00 1988-03-30 12:00:00
## 456 1988-03-31 12:00:00 1988-03-31 12:00:00 1988-03-31 12:00:00
## 457 1988-04-01 12:00:00 1988-04-01 12:00:00 1988-04-01 12:00:00
## 458 1988-04-02 12:00:00 1988-04-02 12:00:00 1988-04-02 12:00:00
## 459 1988-04-03 12:00:00 1988-04-03 12:00:00 1988-04-03 12:00:00
## 460 1988-04-04 12:00:00 1988-04-04 12:00:00 1988-04-04 12:00:00
## 461 1988-04-05 12:00:00 1988-04-05 12:00:00 1988-04-05 12:00:00
## 462 1988-04-06 12:00:00 1988-04-06 12:00:00 1988-04-06 12:00:00
## 463 1988-04-07 12:00:00 1988-04-07 12:00:00 1988-04-07 12:00:00
## 464 1988-04-08 12:00:00 1988-04-08 12:00:00 1988-04-08 12:00:00
## 465 1988-04-09 12:00:00 1988-04-09 12:00:00 1988-04-09 12:00:00
## 466 1988-04-10 12:00:00 1988-04-10 12:00:00 1988-04-10 12:00:00
## 467 1988-04-11 12:00:00 1988-04-11 12:00:00 1988-04-11 12:00:00
## 468 1988-04-12 12:00:00 1988-04-12 12:00:00 1988-04-12 12:00:00
## 469 1988-04-13 12:00:00 1988-04-13 12:00:00 1988-04-13 12:00:00
## 470 1988-04-14 12:00:00 1988-04-14 12:00:00 1988-04-14 12:00:00
## 471 1988-04-15 12:00:00 1988-04-15 12:00:00 1988-04-15 12:00:00
## 472 1988-04-16 12:00:00 1988-04-16 12:00:00 1988-04-16 12:00:00
## 473 1988-04-17 12:00:00 1988-04-17 12:00:00 1988-04-17 12:00:00
## 474 1988-04-18 12:00:00 1988-04-18 12:00:00 1988-04-18 12:00:00
## 475 1988-04-19 12:00:00 1988-04-19 12:00:00 1988-04-19 12:00:00
## 476 1988-04-20 12:00:00 1988-04-20 12:00:00 1988-04-20 12:00:00
## 477 1988-04-21 12:00:00 1988-04-21 12:00:00 1988-04-21 12:00:00
## 478 1988-04-22 12:00:00 1988-04-22 12:00:00 1988-04-22 12:00:00
## 479 1988-04-23 12:00:00 1988-04-23 12:00:00 1988-04-23 12:00:00
## 480 1988-04-24 12:00:00 1988-04-24 12:00:00 1988-04-24 12:00:00
## 481 1988-04-25 12:00:00 1988-04-25 12:00:00 1988-04-25 12:00:00
## 482 1988-04-26 12:00:00 1988-04-26 12:00:00 1988-04-26 12:00:00
## 483 1988-04-27 12:00:00 1988-04-27 12:00:00 1988-04-27 12:00:00
## 484 1988-04-28 12:00:00 1988-04-28 12:00:00 1988-04-28 12:00:00
## 485 1988-04-29 12:00:00 1988-04-29 12:00:00 1988-04-29 12:00:00
## 486 1988-04-30 12:00:00 1988-04-30 12:00:00 1988-04-30 12:00:00
## 487 1988-05-01 12:00:00 1988-05-01 12:00:00 1988-05-01 12:00:00
## 488 1988-05-02 12:00:00 1988-05-02 12:00:00 1988-05-02 12:00:00
## 489 1988-05-03 12:00:00 1988-05-03 12:00:00 1988-05-03 12:00:00
## 490 1988-05-04 12:00:00 1988-05-04 12:00:00 1988-05-04 12:00:00
## 491 1988-05-05 12:00:00 1988-05-05 12:00:00 1988-05-05 12:00:00
## 492 1988-05-06 12:00:00 1988-05-06 12:00:00 1988-05-06 12:00:00
## 493 1988-05-07 12:00:00 1988-05-07 12:00:00 1988-05-07 12:00:00
## 494 1988-05-08 12:00:00 1988-05-08 12:00:00 1988-05-08 12:00:00
## 495 1988-05-09 12:00:00 1988-05-09 12:00:00 1988-05-09 12:00:00
## 496 1988-05-10 12:00:00 1988-05-10 12:00:00 1988-05-10 12:00:00
## 497 1988-05-11 12:00:00 1988-05-11 12:00:00 1988-05-11 12:00:00
## 498 1988-05-12 12:00:00 1988-05-12 12:00:00 1988-05-12 12:00:00
## 499 1988-05-13 12:00:00 1988-05-13 12:00:00 1988-05-13 12:00:00
## 500 1988-05-14 12:00:00 1988-05-14 12:00:00 1988-05-14 12:00:00
## 501 1988-05-15 12:00:00 1988-05-15 12:00:00 1988-05-15 12:00:00
## 502 1988-05-16 12:00:00 1988-05-16 12:00:00 1988-05-16 12:00:00
## 503 1988-05-17 12:00:00 1988-05-17 12:00:00 1988-05-17 12:00:00
## 504 1988-05-18 12:00:00 1988-05-18 12:00:00 1988-05-18 12:00:00
## 505 1988-05-19 12:00:00 1988-05-19 12:00:00 1988-05-19 12:00:00
## 506 1988-05-20 12:00:00 1988-05-20 12:00:00 1988-05-20 12:00:00
## 507 1988-05-21 12:00:00 1988-05-21 12:00:00 1988-05-21 12:00:00
## 508 1988-05-22 12:00:00 1988-05-22 12:00:00 1988-05-22 12:00:00
## 509 1988-05-23 12:00:00 1988-05-23 12:00:00 1988-05-23 12:00:00
## 510 1988-05-24 12:00:00 1988-05-24 12:00:00 1988-05-24 12:00:00
## 511 1988-05-25 12:00:00 1988-05-25 12:00:00 1988-05-25 12:00:00
## 512 1988-05-26 12:00:00 1988-05-26 12:00:00 1988-05-26 12:00:00
## 513 1988-05-27 12:00:00 1988-05-27 12:00:00 1988-05-27 12:00:00
## 514 1988-05-28 12:00:00 1988-05-28 12:00:00 1988-05-28 12:00:00
## 515 1988-05-29 12:00:00 1988-05-29 12:00:00 1988-05-29 12:00:00
## 516 1988-05-30 12:00:00 1988-05-30 12:00:00 1988-05-30 12:00:00
## 517 1988-05-31 12:00:00 1988-05-31 12:00:00 1988-05-31 12:00:00
## 518 1988-06-01 12:00:00 1988-06-01 12:00:00 1988-06-01 12:00:00
## 519 1988-06-02 12:00:00 1988-06-02 12:00:00 1988-06-02 12:00:00
## 520 1988-06-03 12:00:00 1988-06-03 12:00:00 1988-06-03 12:00:00
## 521 1988-06-04 12:00:00 1988-06-04 12:00:00 1988-06-04 12:00:00
## 522 1988-06-05 12:00:00 1988-06-05 12:00:00 1988-06-05 12:00:00
## 523 1988-06-06 12:00:00 1988-06-06 12:00:00 1988-06-06 12:00:00
## 524 1988-06-07 12:00:00 1988-06-07 12:00:00 1988-06-07 12:00:00
## 525 1988-06-08 12:00:00 1988-06-08 12:00:00 1988-06-08 12:00:00
## 526 1988-06-09 12:00:00 1988-06-09 12:00:00 1988-06-09 12:00:00
## 527 1988-06-10 12:00:00 1988-06-10 12:00:00 1988-06-10 12:00:00
## 528 1988-06-11 12:00:00 1988-06-11 12:00:00 1988-06-11 12:00:00
## 529 1988-06-12 12:00:00 1988-06-12 12:00:00 1988-06-12 12:00:00
## 530 1988-06-13 12:00:00 1988-06-13 12:00:00 1988-06-13 12:00:00
## 531 1988-06-14 12:00:00 1988-06-14 12:00:00 1988-06-14 12:00:00
## 532 1988-06-15 12:00:00 1988-06-15 12:00:00 1988-06-15 12:00:00
## 533 1988-06-16 12:00:00 1988-06-16 12:00:00 1988-06-16 12:00:00
## 534 1988-06-17 12:00:00 1988-06-17 12:00:00 1988-06-17 12:00:00
## 535 1988-06-18 12:00:00 1988-06-18 12:00:00 1988-06-18 12:00:00
## 536 1988-06-19 12:00:00 1988-06-19 12:00:00 1988-06-19 12:00:00
## 537 1988-06-20 12:00:00 1988-06-20 12:00:00 1988-06-20 12:00:00
## 538 1988-06-21 12:00:00 1988-06-21 12:00:00 1988-06-21 12:00:00
## 539 1988-06-22 12:00:00 1988-06-22 12:00:00 1988-06-22 12:00:00
## 540 1988-06-23 12:00:00 1988-06-23 12:00:00 1988-06-23 12:00:00
## 541 1988-06-24 12:00:00 1988-06-24 12:00:00 1988-06-24 12:00:00
## 542 1988-06-25 12:00:00 1988-06-25 12:00:00 1988-06-25 12:00:00
## 543 1988-06-26 12:00:00 1988-06-26 12:00:00 1988-06-26 12:00:00
## 544 1988-06-27 12:00:00 1988-06-27 12:00:00 1988-06-27 12:00:00
## 545 1988-06-28 12:00:00 1988-06-28 12:00:00 1988-06-28 12:00:00
## 546 1988-06-29 12:00:00 1988-06-29 12:00:00 1988-06-29 12:00:00
## 547 1988-06-30 12:00:00 1988-06-30 12:00:00 1988-06-30 12:00:00
## 548 1988-07-01 12:00:00 1988-07-01 12:00:00 1988-07-01 12:00:00
## 549 1988-07-02 12:00:00 1988-07-02 12:00:00 1988-07-02 12:00:00
## 550 1988-07-03 12:00:00 1988-07-03 12:00:00 1988-07-03 12:00:00
## 551 1988-07-04 12:00:00 1988-07-04 12:00:00 1988-07-04 12:00:00
## 552 1988-07-05 12:00:00 1988-07-05 12:00:00 1988-07-05 12:00:00
## 553 1988-07-06 12:00:00 1988-07-06 12:00:00 1988-07-06 12:00:00
## 554 1988-07-07 12:00:00 1988-07-07 12:00:00 1988-07-07 12:00:00
## 555 1988-07-08 12:00:00 1988-07-08 12:00:00 1988-07-08 12:00:00
## 556 1988-07-09 12:00:00 1988-07-09 12:00:00 1988-07-09 12:00:00
## 557 1988-07-10 12:00:00 1988-07-10 12:00:00 1988-07-10 12:00:00
## 558 1988-07-11 12:00:00 1988-07-11 12:00:00 1988-07-11 12:00:00
## 559 1988-07-12 12:00:00 1988-07-12 12:00:00 1988-07-12 12:00:00
## 560 1988-07-13 12:00:00 1988-07-13 12:00:00 1988-07-13 12:00:00
## 561 1988-07-14 12:00:00 1988-07-14 12:00:00 1988-07-14 12:00:00
## 562 1988-07-15 12:00:00 1988-07-15 12:00:00 1988-07-15 12:00:00
## 563 1988-07-16 12:00:00 1988-07-16 12:00:00 1988-07-16 12:00:00
## 564 1988-07-17 12:00:00 1988-07-17 12:00:00 1988-07-17 12:00:00
## 565 1988-07-18 12:00:00 1988-07-18 12:00:00 1988-07-18 12:00:00
## 566 1988-07-19 12:00:00 1988-07-19 12:00:00 1988-07-19 12:00:00
## 567 1988-07-20 12:00:00 1988-07-20 12:00:00 1988-07-20 12:00:00
## 568 1988-07-21 12:00:00 1988-07-21 12:00:00 1988-07-21 12:00:00
## 569 1988-07-22 12:00:00 1988-07-22 12:00:00 1988-07-22 12:00:00
## 570 1988-07-23 12:00:00 1988-07-23 12:00:00 1988-07-23 12:00:00
## 571 1988-07-24 12:00:00 1988-07-24 12:00:00 1988-07-24 12:00:00
## 572 1988-07-25 12:00:00 1988-07-25 12:00:00 1988-07-25 12:00:00
## 573 1988-07-26 12:00:00 1988-07-26 12:00:00 1988-07-26 12:00:00
## 574 1988-07-27 12:00:00 1988-07-27 12:00:00 1988-07-27 12:00:00
## 575 1988-07-28 12:00:00 1988-07-28 12:00:00 1988-07-28 12:00:00
## 576 1988-07-29 12:00:00 1988-07-29 12:00:00 1988-07-29 12:00:00
## 577 1988-07-30 12:00:00 1988-07-30 12:00:00 1988-07-30 12:00:00
## 578 1988-07-31 12:00:00 1988-07-31 12:00:00 1988-07-31 12:00:00
## 579 1988-08-01 12:00:00 1988-08-01 12:00:00 1988-08-01 12:00:00
## 580 1988-08-02 12:00:00 1988-08-02 12:00:00 1988-08-02 12:00:00
## 581 1988-08-03 12:00:00 1988-08-03 12:00:00 1988-08-03 12:00:00
## 582 1988-08-04 12:00:00 1988-08-04 12:00:00 1988-08-04 12:00:00
## 583 1988-08-05 12:00:00 1988-08-05 12:00:00 1988-08-05 12:00:00
## 584 1988-08-06 12:00:00 1988-08-06 12:00:00 1988-08-06 12:00:00
## 585 1988-08-07 12:00:00 1988-08-07 12:00:00 1988-08-07 12:00:00
## 586 1988-08-08 12:00:00 1988-08-08 12:00:00 1988-08-08 12:00:00
## 587 1988-08-09 12:00:00 1988-08-09 12:00:00 1988-08-09 12:00:00
## 588 1988-08-10 12:00:00 1988-08-10 12:00:00 1988-08-10 12:00:00
## 589 1988-08-11 12:00:00 1988-08-11 12:00:00 1988-08-11 12:00:00
## 590 1988-08-12 12:00:00 1988-08-12 12:00:00 1988-08-12 12:00:00
## 591 1988-08-13 12:00:00 1988-08-13 12:00:00 1988-08-13 12:00:00
## 592 1988-08-14 12:00:00 1988-08-14 12:00:00 1988-08-14 12:00:00
## 593 1988-08-15 12:00:00 1988-08-15 12:00:00 1988-08-15 12:00:00
## 594 1988-08-16 12:00:00 1988-08-16 12:00:00 1988-08-16 12:00:00
## 595 1988-08-17 12:00:00 1988-08-17 12:00:00 1988-08-17 12:00:00
## 596 1988-08-18 12:00:00 1988-08-18 12:00:00 1988-08-18 12:00:00
## 597 1988-08-19 12:00:00 1988-08-19 12:00:00 1988-08-19 12:00:00
## 598 1988-08-20 12:00:00 1988-08-20 12:00:00 1988-08-20 12:00:00
## 599 1988-08-21 12:00:00 1988-08-21 12:00:00 1988-08-21 12:00:00
## 600 1988-08-22 12:00:00 1988-08-22 12:00:00 1988-08-22 12:00:00
## 601 1988-08-23 12:00:00 1988-08-23 12:00:00 1988-08-23 12:00:00
## 602 1988-08-24 12:00:00 1988-08-24 12:00:00 1988-08-24 12:00:00
## 603 1988-08-25 12:00:00 1988-08-25 12:00:00 1988-08-25 12:00:00
## 604 1988-08-26 12:00:00 1988-08-26 12:00:00 1988-08-26 12:00:00
## 605 1988-08-27 12:00:00 1988-08-27 12:00:00 1988-08-27 12:00:00
## 606 1988-08-28 12:00:00 1988-08-28 12:00:00 1988-08-28 12:00:00
## 607 1988-08-29 12:00:00 1988-08-29 12:00:00 1988-08-29 12:00:00
## 608 1988-08-30 12:00:00 1988-08-30 12:00:00 1988-08-30 12:00:00
## 609 1988-08-31 12:00:00 1988-08-31 12:00:00 1988-08-31 12:00:00
## 610 1988-09-01 12:00:00 1988-09-01 12:00:00 1988-09-01 12:00:00
## 611 1988-09-02 12:00:00 1988-09-02 12:00:00 1988-09-02 12:00:00
## 612 1988-09-03 12:00:00 1988-09-03 12:00:00 1988-09-03 12:00:00
## 613 1988-09-04 12:00:00 1988-09-04 12:00:00 1988-09-04 12:00:00
## 614 1988-09-05 12:00:00 1988-09-05 12:00:00 1988-09-05 12:00:00
## 615 1988-09-06 12:00:00 1988-09-06 12:00:00 1988-09-06 12:00:00
## 616 1988-09-07 12:00:00 1988-09-07 12:00:00 1988-09-07 12:00:00
## 617 1988-09-08 12:00:00 1988-09-08 12:00:00 1988-09-08 12:00:00
## 618 1988-09-09 12:00:00 1988-09-09 12:00:00 1988-09-09 12:00:00
## 619 1988-09-10 12:00:00 1988-09-10 12:00:00 1988-09-10 12:00:00
## 620 1988-09-11 12:00:00 1988-09-11 12:00:00 1988-09-11 12:00:00
## 621 1988-09-12 12:00:00 1988-09-12 12:00:00 1988-09-12 12:00:00
## 622 1988-09-13 12:00:00 1988-09-13 12:00:00 1988-09-13 12:00:00
## 623 1988-09-14 12:00:00 1988-09-14 12:00:00 1988-09-14 12:00:00
## 624 1988-09-15 12:00:00 1988-09-15 12:00:00 1988-09-15 12:00:00
## 625 1988-09-16 12:00:00 1988-09-16 12:00:00 1988-09-16 12:00:00
## 626 1988-09-17 12:00:00 1988-09-17 12:00:00 1988-09-17 12:00:00
## 627 1988-09-18 12:00:00 1988-09-18 12:00:00 1988-09-18 12:00:00
## 628 1988-09-19 12:00:00 1988-09-19 12:00:00 1988-09-19 12:00:00
## 629 1988-09-20 12:00:00 1988-09-20 12:00:00 1988-09-20 12:00:00
## 630 1988-09-21 12:00:00 1988-09-21 12:00:00 1988-09-21 12:00:00
## 631 1988-09-22 12:00:00 1988-09-22 12:00:00 1988-09-22 12:00:00
## 632 1988-09-23 12:00:00 1988-09-23 12:00:00 1988-09-23 12:00:00
## 633 1988-09-24 12:00:00 1988-09-24 12:00:00 1988-09-24 12:00:00
## 634 1988-09-25 12:00:00 1988-09-25 12:00:00 1988-09-25 12:00:00
## 635 1988-09-26 12:00:00 1988-09-26 12:00:00 1988-09-26 12:00:00
## 636 1988-09-27 12:00:00 1988-09-27 12:00:00 1988-09-27 12:00:00
## 637 1988-09-28 12:00:00 1988-09-28 12:00:00 1988-09-28 12:00:00
## 638 1988-09-29 12:00:00 1988-09-29 12:00:00 1988-09-29 12:00:00
## 639 1988-09-30 12:00:00 1988-09-30 12:00:00 1988-09-30 12:00:00
## 640 1988-10-01 12:00:00 1988-10-01 12:00:00 1988-10-01 12:00:00
## 641 1988-10-02 12:00:00 1988-10-02 12:00:00 1988-10-02 12:00:00
## 642 1988-10-03 12:00:00 1988-10-03 12:00:00 1988-10-03 12:00:00
## 643 1988-10-04 12:00:00 1988-10-04 12:00:00 1988-10-04 12:00:00
## 644 1988-10-05 12:00:00 1988-10-05 12:00:00 1988-10-05 12:00:00
## 645 1988-10-06 12:00:00 1988-10-06 12:00:00 1988-10-06 12:00:00
## 646 1988-10-07 12:00:00 1988-10-07 12:00:00 1988-10-07 12:00:00
## 647 1988-10-08 12:00:00 1988-10-08 12:00:00 1988-10-08 12:00:00
## 648 1988-10-09 12:00:00 1988-10-09 12:00:00 1988-10-09 12:00:00
## 649 1988-10-10 12:00:00 1988-10-10 12:00:00 1988-10-10 12:00:00
## 650 1988-10-11 12:00:00 1988-10-11 12:00:00 1988-10-11 12:00:00
## 651 1988-10-12 12:00:00 1988-10-12 12:00:00 1988-10-12 12:00:00
## 652 1988-10-13 12:00:00 1988-10-13 12:00:00 1988-10-13 12:00:00
## 653 1988-10-14 12:00:00 1988-10-14 12:00:00 1988-10-14 12:00:00
## 654 1988-10-15 12:00:00 1988-10-15 12:00:00 1988-10-15 12:00:00
## 655 1988-10-16 12:00:00 1988-10-16 12:00:00 1988-10-16 12:00:00
## 656 1988-10-17 12:00:00 1988-10-17 12:00:00 1988-10-17 12:00:00
## 657 1988-10-18 12:00:00 1988-10-18 12:00:00 1988-10-18 12:00:00
## 658 1988-10-19 12:00:00 1988-10-19 12:00:00 1988-10-19 12:00:00
## 659 1988-10-20 12:00:00 1988-10-20 12:00:00 1988-10-20 12:00:00
## 660 1988-10-21 12:00:00 1988-10-21 12:00:00 1988-10-21 12:00:00
## 661 1988-10-22 12:00:00 1988-10-22 12:00:00 1988-10-22 12:00:00
## 662 1988-10-23 12:00:00 1988-10-23 12:00:00 1988-10-23 12:00:00
## 663 1988-10-24 12:00:00 1988-10-24 12:00:00 1988-10-24 12:00:00
## 664 1988-10-25 12:00:00 1988-10-25 12:00:00 1988-10-25 12:00:00
## 665 1988-10-26 12:00:00 1988-10-26 12:00:00 1988-10-26 12:00:00
## 666 1988-10-27 12:00:00 1988-10-27 12:00:00 1988-10-27 12:00:00
## 667 1988-10-28 12:00:00 1988-10-28 12:00:00 1988-10-28 12:00:00
## 668 1988-10-29 12:00:00 1988-10-29 12:00:00 1988-10-29 12:00:00
## 669 1988-10-30 12:00:00 1988-10-30 12:00:00 1988-10-30 12:00:00
## 670 1988-10-31 12:00:00 1988-10-31 12:00:00 1988-10-31 12:00:00
## 671 1988-11-01 12:00:00 1988-11-01 12:00:00 1988-11-01 12:00:00
## 672 1988-11-02 12:00:00 1988-11-02 12:00:00 1988-11-02 12:00:00
## 673 1988-11-03 12:00:00 1988-11-03 12:00:00 1988-11-03 12:00:00
## 674 1988-11-04 12:00:00 1988-11-04 12:00:00 1988-11-04 12:00:00
## 675 1988-11-05 12:00:00 1988-11-05 12:00:00 1988-11-05 12:00:00
## 676 1988-11-06 12:00:00 1988-11-06 12:00:00 1988-11-06 12:00:00
## 677 1988-11-07 12:00:00 1988-11-07 12:00:00 1988-11-07 12:00:00
## 678 1988-11-08 12:00:00 1988-11-08 12:00:00 1988-11-08 12:00:00
## 679 1988-11-09 12:00:00 1988-11-09 12:00:00 1988-11-09 12:00:00
## 680 1988-11-10 12:00:00 1988-11-10 12:00:00 1988-11-10 12:00:00
## 681 1988-11-11 12:00:00 1988-11-11 12:00:00 1988-11-11 12:00:00
## 682 1988-11-12 12:00:00 1988-11-12 12:00:00 1988-11-12 12:00:00
## 683 1988-11-13 12:00:00 1988-11-13 12:00:00 1988-11-13 12:00:00
## 684 1988-11-14 12:00:00 1988-11-14 12:00:00 1988-11-14 12:00:00
## 685 1988-11-15 12:00:00 1988-11-15 12:00:00 1988-11-15 12:00:00
## 686 1988-11-16 12:00:00 1988-11-16 12:00:00 1988-11-16 12:00:00
## 687 1988-11-17 12:00:00 1988-11-17 12:00:00 1988-11-17 12:00:00
## 688 1988-11-18 12:00:00 1988-11-18 12:00:00 1988-11-18 12:00:00
## 689 1988-11-19 12:00:00 1988-11-19 12:00:00 1988-11-19 12:00:00
## 690 1988-11-20 12:00:00 1988-11-20 12:00:00 1988-11-20 12:00:00
## 691 1988-11-21 12:00:00 1988-11-21 12:00:00 1988-11-21 12:00:00
## 692 1988-11-22 12:00:00 1988-11-22 12:00:00 1988-11-22 12:00:00
## 693 1988-11-23 12:00:00 1988-11-23 12:00:00 1988-11-23 12:00:00
## 694 1988-11-24 12:00:00 1988-11-24 12:00:00 1988-11-24 12:00:00
## 695 1988-11-25 12:00:00 1988-11-25 12:00:00 1988-11-25 12:00:00
## 696 1988-11-26 12:00:00 1988-11-26 12:00:00 1988-11-26 12:00:00
## 697 1988-11-27 12:00:00 1988-11-27 12:00:00 1988-11-27 12:00:00
## 698 1988-11-28 12:00:00 1988-11-28 12:00:00 1988-11-28 12:00:00
## 699 1988-11-29 12:00:00 1988-11-29 12:00:00 1988-11-29 12:00:00
## 700 1988-11-30 12:00:00 1988-11-30 12:00:00 1988-11-30 12:00:00
## 701 1988-12-01 12:00:00 1988-12-01 12:00:00 1988-12-01 12:00:00
## 702 1988-12-02 12:00:00 1988-12-02 12:00:00 1988-12-02 12:00:00
## 703 1988-12-03 12:00:00 1988-12-03 12:00:00 1988-12-03 12:00:00
## 704 1988-12-04 12:00:00 1988-12-04 12:00:00 1988-12-04 12:00:00
## 705 1988-12-05 12:00:00 1988-12-05 12:00:00 1988-12-05 12:00:00
## 706 1988-12-06 12:00:00 1988-12-06 12:00:00 1988-12-06 12:00:00
## 707 1988-12-07 12:00:00 1988-12-07 12:00:00 1988-12-07 12:00:00
## 708 1988-12-08 12:00:00 1988-12-08 12:00:00 1988-12-08 12:00:00
## 709 1988-12-09 12:00:00 1988-12-09 12:00:00 1988-12-09 12:00:00
## 710 1988-12-10 12:00:00 1988-12-10 12:00:00 1988-12-10 12:00:00
## 711 1988-12-11 12:00:00 1988-12-11 12:00:00 1988-12-11 12:00:00
## 712 1988-12-12 12:00:00 1988-12-12 12:00:00 1988-12-12 12:00:00
## 713 1988-12-13 12:00:00 1988-12-13 12:00:00 1988-12-13 12:00:00
## 714 1988-12-14 12:00:00 1988-12-14 12:00:00 1988-12-14 12:00:00
## 715 1988-12-15 12:00:00 1988-12-15 12:00:00 1988-12-15 12:00:00
## 716 1988-12-16 12:00:00 1988-12-16 12:00:00 1988-12-16 12:00:00
## 717 1988-12-17 12:00:00 1988-12-17 12:00:00 1988-12-17 12:00:00
## 718 1988-12-18 12:00:00 1988-12-18 12:00:00 1988-12-18 12:00:00
## 719 1988-12-19 12:00:00 1988-12-19 12:00:00 1988-12-19 12:00:00
## 720 1988-12-20 12:00:00 1988-12-20 12:00:00 1988-12-20 12:00:00
## 721 1988-12-21 12:00:00 1988-12-21 12:00:00 1988-12-21 12:00:00
## 722 1988-12-22 12:00:00 1988-12-22 12:00:00 1988-12-22 12:00:00
## 723 1988-12-23 12:00:00 1988-12-23 12:00:00 1988-12-23 12:00:00
## 724 1988-12-24 12:00:00 1988-12-24 12:00:00 1988-12-24 12:00:00
## 725 1988-12-25 12:00:00 1988-12-25 12:00:00 1988-12-25 12:00:00
## 726 1988-12-26 12:00:00 1988-12-26 12:00:00 1988-12-26 12:00:00
## 727 1988-12-27 12:00:00 1988-12-27 12:00:00 1988-12-27 12:00:00
## 728 1988-12-28 12:00:00 1988-12-28 12:00:00 1988-12-28 12:00:00
## 729 1988-12-29 12:00:00 1988-12-29 12:00:00 1988-12-29 12:00:00
## 730 1988-12-30 12:00:00 1988-12-30 12:00:00 1988-12-30 12:00:00
## 731 1988-12-31 12:00:00 1988-12-31 12:00:00 1988-12-31 12:00:00

# Maximum difference (small numerical differences)
unlist( lapply(
    X      = colnames(tmp1), 
    FUN    = function(X){ 
        max( tmp1.c[,X] - tmp1.b[,X] ) 
    }   
) ) 

## [1] -1  0  0  0  0  0

#   NB: difference in time zone! + real difference in time.

#   Now compare the bin file themselves
tmp1a <- readBin( con = filenm, what = "raw", n = 1000000 ) 
tmp2a <- readBin( con = filenm2, what = "raw", n = 1000000 ) 

#   FALSE: There is a difference in the min - max after the column name
#   + eventual differences in time if not 12:00 
identical( tmp1a, tmp2a ) 

## [1] FALSE

test <- tmp1a != tmp2a 
readBin( con = tmp1a[ test ], what = "double", size = 4, n = sum(test) ) 

## [1] -4.800000e+00  2.718399e+23  4.185255e-08  1.261292e-04  4.657850e-16
## [6] -1.531228e+14  1.905526e+17  2.713045e-35 -3.009684e-03

readBin( con = tmp2a[ test ], what = "double", size = 4, n = sum(test) ) 

## [1] 0 0 0 0 0 0 0 0 0

# Remove the temporary file
file.remove( filenm ) 

## [1] TRUE

file.remove( filenm2 ) 

## [1] TRUE

[Package macroutils version 1.14.0 Index]